+++ CHANGELOG.md
@@ -19,6 +19,8 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 - [PR-89](https://github.com/itk-dev/book_aarhus/pull/89): Fixed issue with GET /user-bookings where pagination
 was not handled correctly.
+- [PR-90](https://github.com/itk-dev/book_aarhus/pull/90): Updated fixtures to match expected data for resource
+categories.
 
 ## [1.0.0] - YYYY-MM-DD